Add support for launching with libvirt provider
In order to be able to use vagrant with libvirt we need to set
a few configuration options.

This patch requires either to provide a libvirt compatible vagrant
image or to add instructions in the documentation in order to
provide the required steps to convert a virtualbox image into
a libvirt-compatible one.